[add] delete vector class in DynamicsMinuteChange

+++ src/quabla/simulator/dynamics/DynamicsTipOff.java
@@ -132,10 +132,10 @@ public class DynamicsTipOff extends AbstractDynamics {
 		MathematicalVector quatdot = tensor.dot(quat).multiply(0.5);
 
 		DynamicsMinuteChangeTrajectory delta = new DynamicsMinuteChangeTrajectory();
-		delta.setDeltaPosNED(velENU);
-		delta.setDeltaVelNED(accENU);
-		delta.setDeltaOmegaBODY(omegadot);
-		delta.setDeltaQuat(quatdot);
+		delta.setDeltaPosNED(velENU.toDouble());
+		delta.setDeltaVelNED(accENU.toDouble());
+		delta.setDeltaOmegaBODY(omegadot.toDouble());
+		delta.setDeltaQuat(quatdot.toDouble());
 
 		return delta;
 	}
